bugfix: move an api to login with swagger
