+++ aspnet-core/src/Government.Services.Application.Contracts/Authorization/Accounts/IAccountAppService.cs
@@ -6,8 +6,6 @@ namespace Government.Services.Authorization.Accounts
 {
     public interface IAccountAppService : IApplicationService
     {
-        Task<IsTenantAvailableOutput> IsTenantAvailable(IsTenantAvailableInput input);
-
         Task<RegisterOutput> Register(RegisterInput input);
     }
 }
