+++ aspnet-core/src/Government.Services.Application/Authorization/Accounts/AccountAppService.cs
@@ -21,22 +21,6 @@ namespace Government.Services.Authorization.Accounts
             _userRegistrationManager = userRegistrationManager;
         }
 
-        public async Task<IsTenantAvailableOutput> IsTenantAvailable(IsTenantAvailableInput input)
-        {
-            var tenant = await TenantManager.FindByTenancyNameAsync(input.TenancyName);
-            if (tenant == null)
-            {
-                return new IsTenantAvailableOutput(TenantAvailabilityState.NotFound);
-            }
-
-            if (!tenant.IsActive)
-            {
-                return new IsTenantAvailableOutput(TenantAvailabilityState.InActive);
-            }
-
-            return new IsTenantAvailableOutput(TenantAvailabilityState.Available, tenant.Id);
-        }
-
         public async Task<RegisterOutput> Register(RegisterInput input)
         {
             var user = await _userRegistrationManager.RegisterAsync(
