+++ aspnet-core/src/Government.Services.Web.Host/wwwroot/swagger/ui/abp.swagger.js
@@ -91,7 +91,7 @@ var abp = abp || {};
                 }
             };
 
-            xhrTenancyName.open('POST', '/api/services/app/Account/IsTenantAvailable', true);
+            xhrTenancyName.open('POST', '/api/TokenAuth/IsTenantAvailable?', true);
             xhrTenancyName.setRequestHeader('Content-type', 'application/json');
             addAntiForgeryTokenToXhr(xhrTenancyName);
             xhrTenancyName.send(
